대여목록 반납상태 변경 설정중

+++ src/main/java/com/itwillbs/bookjuk/repository/RentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.itwillbs.bookjuk.entity.RentEntity;
 
-public interface RentRepository extends JpaRepository<RentEntity, Long> {
+public interface RentRepository extends JpaRepository<RentEntity, Integer> {
 
 }
